Fix Spacing

+++ onnxruntime/core/mlas/lib/qgemm.cpp
@@ -2245,7 +2245,7 @@ Return Value:
     int32_t TargetThreadCount;
 
     if (Complexity < double(MLAS_QGEMM_THREAD_COMPLEXITY * MlasPlatform.MaximumThreadCount)) {
-        TargetThreadCount = int32_t(Complexity / MLAS_QGEMM_THREAD_COMPLEXITY) + 1;
+        TargetThreadCount = int32_t(Complexity / double(MLAS_QGEMM_THREAD_COMPLEXITY)) + 1;
     } else {
         TargetThreadCount = MlasPlatform.MaximumThreadCount;
     }
